requirements-dev.txt
# ================================================
# DEPENDENCIAS PARA DESARROLLO LOCAL
# ================================================
# Este archivo NO incluye psycopg2-binary (PostgreSQL)
# En desarrollo local se usa SQLite

# Django Framework
Django==5.2.7

# Google Sheets Integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# PDF Processing
PyMuPDF>=1.24.0

# Environment variables
python-dotenv==1.0.0

# Database utilities (necesario para parsing URLs)
dj-database-url==2.1.0

# Static files (útil también en desarrollo)
whitenoise==6.6.0
